Start lexer code to read utf-8 chars from potentially long line
We take this approach, mostly borrowed from here

  https://stackoverflow.com/questions/37079342/

because we want to be able to handle a large file with potentially
only a single line, without loading the entire line/file into memory
at once. This also gives us iteration over utf-8 chars, which is
definitely what we want.

TODO: Add some code to detect file encoding and error if not utf-8.

Later we could handle other encodings, maybe, but we're definitely
not there yet.