Implement basic scanner for one-char tokens
We are following along with Robert Nystrom's _Crafting Interpreters_
book. The plan is to implement a version of his interpreter in Rust,
and then modify it a bit to implement our own language, to get some
hands-on experience. This is the first part of his scanner, in Rust.

We store the current token and the two next tokens, which may be None
in our scanner, to allow two-character lookahead. This works well with
how we've implemented our file utf-8 reader.

Next we will implement more complex token types.